Factor out OutboundLinkButton
